refactor: cleanup, WIP
