+++ lib/api/api-request.js
@@ -5,7 +5,6 @@ const assert = require('node:assert')
 
 const { InvalidArgumentError, RequestAbortedError } = require('../core/errors')
 const util = require('../core/util')
-const RedirectHandler = require('../handler/redirect-handler')
 
 const { Readable } = require('./readable')
 const { getResolveErrorBodyCallback } = require('./util')
@@ -193,15 +192,6 @@ function request (opts, callback) {
 
   try {
     const handler = new RequestHandler(opts, callback)
-
-    if (opts?.maxRedirections != null && (!Number.isInteger(opts?.maxRedirections) || opts?.maxRedirections < 0)) {
-      throw new InvalidArgumentError('maxRedirections must be a positive number')
-    }
-
-    if (opts?.maxRedirections > 0) {
-      RedirectHandler.buildDispatch(this, opts.maxRedirections)(opts, handler)
-      return
-    }
 
     this.dispatch(opts, handler)
   } catch (err) {
