+++ lib/dispatcher/client.js
@@ -49,7 +49,6 @@ const {
   kClose,
   kDestroy,
   kDispatch,
-  kInterceptors,
   kLocalAddress,
   kMaxResponseSize,
   kOnError,
@@ -59,7 +58,6 @@ const {
 } = require('../core/symbols.js')
 const connectH1 = require('./client-h1.js')
 const connectH2 = require('./client-h2.js')
-let deprecatedInterceptorWarned = false
 
 const kClosedResolve = Symbol('kClosedResolve')
 
@@ -77,7 +75,6 @@ class Client extends DispatcherBase {
    * @param {import('../../types/client.js').Client.Options} options
    */
   constructor (url, {
-    interceptors,
     maxHeaderSize,
     headersTimeout,
     socketTimeout,
@@ -206,18 +203,6 @@ class Client extends DispatcherBase {
         ...(autoSelectFamily ? { autoSelectFamily, autoSelectFamilyAttemptTimeout } : undefined),
         ...connect
       })
-    }
-
-    if (interceptors?.Client && Array.isArray(interceptors.Client)) {
-      this[kInterceptors] = interceptors.Client
-      if (!deprecatedInterceptorWarned) {
-        deprecatedInterceptorWarned = true
-        process.emitWarning('Client.Options#interceptor is deprecated. Use Dispatcher#compose instead.', {
-          code: 'UNDICI-CLIENT-INTERCEPTOR-DEPRECATED'
-        })
-      }
-    } else {
-      this[kInterceptors] = [createRedirectInterceptor({ maxRedirections })]
     }
 
     this[kUrl] = util.parseOrigin(url)
@@ -361,8 +346,6 @@ class Client extends DispatcherBase {
     })
   }
 }
-
-const createRedirectInterceptor = require('../interceptor/redirect-interceptor.js')
 
 function onError (client, err) {
   if (
